Modified Invoker, folderpollingserver, Androiddispatcher
Now from OSGI the engineer can test by: tca event start_test <testset>
[testcase1]...[testcasex]

Now AndroidDispatcher sends an action (like before) but also waits until
the UE finished testing. After that, AndroidDispatcher receives the
results and saves it to a XML file.

Now there is an serverinit.ini file where the PATH to the sestscripts
and to the results directory are stored. If this file can't be found,
than the server takes a default directory. (on this moment, i don't know
in witch directory i must store this .ini file..